+++ components/model-builder/config.js
@@ -1,0 +1,9 @@
+export default {
+	defaultScale: 1.3,
+	blockWidth: {
+		compactParams: 110,
+		compact: 60,
+		normalParams: 220,
+		normal: 110
+	}
+};